# small perf optimizing. Only remove whiteout path if it needs to be included in base image